backend/pyproject.toml
[project]
name = "structuredmatch-backend"
version = "0.1.0"
description = "Asset-manager copilot for structured products — Phase 1-5 backend MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "faiss-cpu>=1.9",
    "numpy>=2.0",
    "google-genai>=0.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "httpx>=0.27",
    "ruff>=0.7",
    "mypy>=1.13",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B"]
# B008: Depends(...) in argument defaults is the canonical FastAPI pattern.
# E501: line length is governed by the formatter.
ignore = ["E501", "B008"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict_optional = true
exclude = ["app/tests/"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
